requirements.txt
flask==2.2.5
flask-login==0.6.3
email-validator==2.0.0.post2
flask-wtf==1.1.1
flask_sqlalchemy==3.0.5
requests==2.31.0
pydantic==2.5.3
flask-socketio==5.3.4
python-socketio==5.7.2
eventlet==0.38.2
flask-migrate==4.1.0
redis==4.3.5
flask-limiter==3.5.0
flask-talisman==1.1.0
python-dotenv==0.19.2
pytest==7.4.4